refactor: Remove usages of Seam's Work class
Turned the seam class into a functional utility.
This revision still performs the same code as the Seam Work class.
