+++ zanata-war/src/main/java/org/zanata/service/impl/CopyTransServiceImpl.java
@@ -43,6 +43,7 @@ import org.zanata.model.HTextFlow;
 import org.zanata.service.CopyTransService;
 import org.zanata.service.LocaleService;
 import org.zanata.service.TranslationStateCache;
+import org.zanata.transaction.TransactionUtil;
 import org.zanata.util.ServiceLocator;
 import com.google.common.base.Optional;
 import com.google.common.base.Stopwatch;
@@ -52,6 +53,8 @@ import lombok.NoArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
 import javax.validation.constraints.NotNull;
+
+import static org.zanata.transaction.TransactionUtil.runInTransaction;
 
 @Name("copyTransServiceImpl")
 @Scope(ScopeType.STATELESS)
@@ -186,10 +189,12 @@ public class CopyTransServiceImpl implements CopyTransService {
             List<HTextFlow> copyTargets =
                     docTextFlows.subList(batchStart, batchEnd);
             Integer numCopied =
-                    copyTransWorkFactory.createCopyTransWork(targetLocale,
-                            options, document, requireTranslationReview,
-                            copyTargets)
-                            .workInTransaction();
+                    runInTransaction(
+                            copyTransWorkFactory
+                                    .createCopyTransExecution(targetLocale,
+                                            options, document,
+                                            requireTranslationReview,
+                                            copyTargets));
             if (taskHandleOpt.isPresent()) {
                 taskHandleOpt.get().increaseProgress(batchSize);
             }
