+++ zanata-war/src/main/java/org/zanata/service/impl/CopyVersionServiceImpl.java
@@ -1,9 +1,17 @@
 package org.zanata.service.impl;
 
-import com.google.common.base.Optional;
-import com.google.common.base.Stopwatch;
-import com.google.common.collect.Lists;
+import static org.zanata.transaction.TransactionUtil.runInTransaction;
+
+import java.util.Collections;
+import java.util.List;
+import java.util.Map;
+import java.util.concurrent.Future;
+
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+
 import lombok.extern.slf4j.Slf4j;
+
 import org.jboss.seam.ScopeType;
 import org.jboss.seam.annotations.AutoCreate;
 import org.jboss.seam.annotations.In;
@@ -37,11 +45,10 @@ import org.zanata.service.VersionStateCache;
 import org.zanata.util.JPACopier;
 import org.zanata.util.TranslationUtil;
 
-import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
-import java.util.Collections;
-import java.util.Map;
-import java.util.concurrent.Future;
+import com.google.common.base.Optional;
+import com.google.common.base.Stopwatch;
+import com.google.common.collect.Lists;
+import com.google.common.collect.Maps;
 
 /**
  * @author Alex Eng <a href="mailto:aeng@redhat.com">aeng@redhat.com</a>
@@ -189,13 +196,54 @@ public class CopyVersionServiceImpl implements CopyVersionService {
     protected Map<Long, Long> copyDocumentBatch(Long versionId,
             Long newVersionId, int batchStart, int batchLength) {
         try {
-            return new CopyDocumentWork(versionId, newVersionId, documentDAO,
-                    projectIterationDAO, filePersistService, this,
-                    batchStart, batchLength).workInTransaction();
+            return runInTransaction(() -> copyDocument(
+                    versionId, newVersionId, batchStart, batchLength));
         } catch (Exception e) {
             log.warn("exception during copy document", e);
             return Collections.emptyMap();
         }
+    }
+
+    /**
+     * Copy documents from HProjectIteration(id=versionId) in batches(batchStart,
+     * batchLength) into HProjectIteration(id=newVersionId). Copying includes
+     * HRawDocument stored in file system.
+     * @return Map indexed by the original document id, and to the new copied
+     * document id.
+     * @throws Exception
+     */
+    private Map<Long, Long> copyDocument(final Long versionId,
+            final Long newVersionId,
+            final int batchStart,
+            final int batchLength) throws Exception {
+        Map<Long, Long> docMap = Maps.newHashMap();
+
+        HProjectIteration newVersion =
+                projectIterationDAO.findById(newVersionId);
+
+        List<HDocument> documents = documentDAO.findAllByVersionId(versionId,
+                batchStart, batchLength);
+
+        for (HDocument doc : documents) {
+            HDocument newDocument = copyDocument(newVersion, doc);
+            // Needs to persist before inserting raw document
+            newDocument = documentDAO.makePersistent(newDocument);
+
+            if (doc.getRawDocument() != null) {
+                HRawDocument newRawDocument =
+                        copyRawDocument(newDocument, doc.getRawDocument());
+
+                filePersistService.copyAndPersistRawDocument(
+                        doc.getRawDocument(), newRawDocument);
+
+                documentDAO.addRawDocument(newDocument, newRawDocument);
+            }
+            newVersion.getDocuments()
+                    .put(newDocument.getDocId(), newDocument);
+            docMap.put(doc.getId(), newDocument.getId());
+        }
+        documentDAO.flush();
+        return docMap;
     }
 
     /**
@@ -252,13 +300,39 @@ public class CopyVersionServiceImpl implements CopyVersionService {
     protected Map<Long, Long> copyTextFlowBatch(Long documentId,
             Long newDocumentId, int batchStart, int batchLength) {
         try {
-            return new CopyTextFlowWork(documentId, newDocumentId, textFlowDAO,
-                    documentDAO, this, batchStart, batchLength)
-                    .workInTransaction();
+            return runInTransaction(() -> this.copyTextFlows(documentId,
+                    newDocumentId, batchStart, batchLength));
         } catch (Exception e) {
             log.warn("exception during copy text flow", e);
             return Collections.EMPTY_MAP;
         }
+    }
+
+    private Map<Long, Long> copyTextFlows(
+            final Long documentId,
+            final Long newDocumentId,
+            final int batchStart,
+            final int batchLength) throws Exception {
+
+        Map<Long, HTextFlow> tfMap = Maps.newHashMap();
+
+        List<HTextFlow> textFlows = textFlowDAO.getTextFlowsByDocumentId(
+                documentId, batchStart, batchLength);
+
+        HDocument newDocument = documentDAO.getById(newDocumentId);
+        for (HTextFlow textFlow : textFlows) {
+            HTextFlow newTextFlow =
+                    copyTextFlow(newDocument, textFlow);
+
+            newDocument.getTextFlows().add(newTextFlow);
+            newDocument.getAllTextFlows()
+                    .put(newTextFlow.getResId(), newTextFlow);
+            tfMap.put(textFlow.getId(), newTextFlow);
+        }
+        documentDAO.makePersistent(newDocument);
+        documentDAO.flush();
+
+        return Maps.transformEntries(tfMap, (key, value) -> value.getId());
     }
 
     /**
@@ -272,13 +346,41 @@ public class CopyVersionServiceImpl implements CopyVersionService {
     protected int copyTextFlowTargetBatch(Long tfId, Long newTfId,
             int batchStart, int batchLength) {
         try {
-            return new CopyTextFlowTargetWork(tfId, newTfId, textFlowTargetDAO,
-                    textFlowDAO, this, batchStart, batchLength)
-                    .workInTransaction();
+            return runInTransaction(() -> this.copyTextFlowTargets(tfId,
+                    newTfId, batchStart, batchLength));
         } catch (Exception e) {
             log.warn("exception during copy text flow target", e);
             return 0;
         }
+    }
+
+    /**
+     * Copy HTextFlowTarget from HTextFlow(id=tfId) in batches(batchStart,
+     * batchLength) into HTextFlow(id=newTfId).
+     * @return Number of text flow targets copied.
+     * @throws Exception
+     */
+    private Integer copyTextFlowTargets(final Long tfId,
+            final Long newTfId,
+            final int batchStart,
+            final int batchLength) throws Exception {
+        HTextFlow newTextFlow = textFlowDAO.findById(newTfId);
+
+        List<HTextFlowTarget> copyTargets =
+                textFlowTargetDAO
+                        .getByTextFlowId(tfId, batchStart, batchLength);
+
+        for (HTextFlowTarget tft : copyTargets) {
+            HTextFlowTarget newTextFlowTarget =
+                    copyTextFlowTarget(newTextFlow, tft);
+
+            newTextFlow.getTargets()
+                    .put(newTextFlowTarget.getLocale().getId(),
+                            newTextFlowTarget);
+        }
+        textFlowDAO.makePersistent(newTextFlow);
+        textFlowDAO.flush();
+        return copyTargets.size();
     }
 
     @Override
